Basic Autonomous Corrections
Fixed errors in basic autonomous (Removed all code from While Loops).
